[Refactor] #55 봉사활동 동기화 서비스를 인터페이스와 구현체로 분리
